+++ tabletest-junit/src/main/java/io/github/nchaugen/tabletest/renderer/MarkdownRenderer.java
@@ -17,6 +17,7 @@ package io.github.nchaugen.tabletest.renderer;
 
 import io.github.nchaugen.tabletest.parser.Row;
 import io.github.nchaugen.tabletest.parser.Table;
+import org.junit.jupiter.api.extension.ExtensionContext;
 
 import java.util.Collection;
 import java.util.List;
@@ -27,9 +28,14 @@ import java.util.stream.Collectors;
 
 public class MarkdownRenderer implements TableRenderer {
 
+    private static final String NEWLINE = "\n";
+
     @Override
-    public String render(Table table, ColumnRoles columnRoles) {
-        return String.join("\n",
+    public String render(Table table, ColumnRoles columnRoles, ExtensionContext context) {
+        return String.join(
+            NEWLINE,
+            "## " + context.getDisplayName(),
+            "",
             table.headers().stream().map(this::render).collect(MARKDOWN_ROW),
             table.headers().stream().map(__ -> "---").collect(MARKDOWN_ROW),
             table.rows().stream().map(this::render).collect(MULTILINE)
@@ -40,7 +46,7 @@ public class MarkdownRenderer implements TableRenderer {
         MARKDOWN_ROW = Collectors.joining(" | ", "| ", " |");
 
     private static final Collector<CharSequence, ?, String>
-        MULTILINE = Collectors.joining("\n", "", "\n");
+        MULTILINE = Collectors.joining(NEWLINE, "", NEWLINE);
 
     private String render(Row row) {
         return row.values().stream().map(this::render).collect(MARKDOWN_ROW);