+++ tabletest-junit/src/test/java/io/github/nchaugen/tabletest/renderer/MarkdownRendererTest.java
@@ -2,18 +2,22 @@ package io.github.nchaugen.tabletest.renderer;
 
 import io.github.nchaugen.tabletest.parser.TableParser;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtensionContext;
 
 import static io.github.nchaugen.tabletest.renderer.ColumnRoles.NO_ROLES;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 public class MarkdownRendererTest {
 
+    private final ExtensionContext context = new StubExtensionContext();
     private final MarkdownRenderer renderer = new MarkdownRenderer();
 
     @Test
     void shouldRenderBasicTable() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | c |
                 | --- | --- | --- |
                 | 1 | 2 | 3 |
@@ -23,7 +27,8 @@ public class MarkdownRendererTest {
                 a | b | c
                 1 | 2 | 3
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }
@@ -32,6 +37,8 @@ public class MarkdownRendererTest {
     void shouldRenderNull() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | c |
                 | --- | --- | --- |
                 | 1 |  | 3 |
@@ -41,7 +48,8 @@ public class MarkdownRendererTest {
                 a | b | c
                 1 |   | 3
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }
@@ -50,6 +58,8 @@ public class MarkdownRendererTest {
     void shouldRenderTableWithEscapedPipe() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | a\\|b |
                 | --- | --- | --- |
                 | \\| | \\| | Text with \\| character |
@@ -59,7 +69,8 @@ public class MarkdownRendererTest {
                 a   | b   | 'a|b'
                 "|" | '|' | "Text with | character"
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }
@@ -68,6 +79,8 @@ public class MarkdownRendererTest {
     void shouldRenderTableWithList() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | c |
                 | --- | --- | --- |
                 | [] | [1, 2, 3] | [\\|, \\|] |
@@ -77,7 +90,8 @@ public class MarkdownRendererTest {
                 a  | b         | c
                 [] | [1,2,3] | ['|', "|"]
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }
@@ -86,6 +100,8 @@ public class MarkdownRendererTest {
     void shouldRenderTableWithSet() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | c |
                 | --- | --- | --- |
                 | {} | {1, 2, 3} | {\\|\\|} |
@@ -95,7 +111,8 @@ public class MarkdownRendererTest {
                 a  | b   | c
                 {} | {1,2,3} | {"||"}
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }
@@ -104,6 +121,8 @@ public class MarkdownRendererTest {
     void shouldRenderTableWithMap() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | c |
                 | --- | --- | --- |
                 | [:] | [a: 1, b: 2, c: 3] | [b: \\|\\|] |
@@ -113,7 +132,8 @@ public class MarkdownRendererTest {
                 a   | b             | c
                 [:] | [a:1,b:2,c:3] | [b: "||"]
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }
@@ -122,6 +142,8 @@ public class MarkdownRendererTest {
     void shouldRenderTableWithNestedList() {
         assertEquals(
             """
+                ## Display Name
+                
                 | a | b | c |
                 | --- | --- | --- |
                 | [{}] | [[1], [2, 3]] | [[a: \\|], [b: [\\|]]] |
@@ -131,7 +153,8 @@ public class MarkdownRendererTest {
                 a    | b           | c
                 [{}] | [[1],[2,3]] | [[a: '|'], [b: ["|"]]]
                 """),
-                NO_ROLES
+                NO_ROLES,
+                context
             )
         );
     }